LayoutTests/platform/chromium-mac/fast/notifications/notifications-document-close-crash-expected.txt

CONSOLE MESSAGE: line 36: Uncaught Error: InvalidStateError: DOM Exception 11
PASS
